requirements.txt
# Database
pgdbtoolkit==1.0.0

# Redis
redis==5.2.0

# API
fastapi==0.104.1
uvicorn==0.24.0
slowapi==0.1.9

# Pydantic
pydantic==2.4.2
pydantic-settings==2.7.1
pydantic[email]

# JWT
python-jose[cryptography]==3.3.0
PyJWT==2.8.0

# OAuth2 y archivos
python-multipart==0.0.12

# Password hashing
bcrypt==4.0.1
passlib[bcrypt]==1.7.4

# Sendgrid
sendgrid==6.11.0

# Transbank
transbank-sdk==4.0.0

# Dotenv
python-dotenv==1.0.0

# HTTPX
httpx==0.25.1